WIP: removed alignment streams (will not work)
The AIE memory is to small to load a window with the whole alignment, so
need for reusing the kernel.

Added some code to the hls kernels to fill the alignments up with zeroes
to fill a window if it is not fully used.

Now need for a way to keep the EV and branch lenghts matrices inside the kernel
for as long as needed and replaced for the next, like a coefficient
array.