Add window2in transpose on read + other optimizations
Now the branch and ev matrices are only read once and from PL memory
resend to the AIE. Also the transpose is now in a seperate file and
compiled into the project. The axiu<128,0,0,0> x variable is now not
defined at the top of the function, but per scope of a for loop, in the
hope that the compiler is not hindered in optimizing it as it might be
scared of a data dependency that is not there. This is also done for the
ap_uint<512> buffer variable.